+++ Watchman.Discord/Areas/Protection/Controllers/MuteUserController.cs
@@ -12,7 +12,7 @@ using Devscord.DiscordFramework.Services.Factories;
 using Watchman.Discord.Areas.Protection.Commands;
 using Watchman.Discord.Areas.Protection.Models;
 using Watchman.Discord.Areas.Protection.Services;
-using Watchman.DomainModel.Users;
+using Watchman.DomainModel.Protection;
 
 namespace Watchman.Discord.Areas.Protection.Controllers
 {
