+++ Watchman.Discord/Areas/Protection/Services/MutingHelper.cs
@@ -5,9 +5,9 @@ using Devscord.DiscordFramework.Commons.Exceptions;
 using Devscord.DiscordFramework.Middlewares.Contexts;
 using Devscord.DiscordFramework.Services;
 using Watchman.Cqrs;
-using Watchman.DomainModel.Users;
-using Watchman.DomainModel.Users.Commands;
-using Watchman.DomainModel.Users.Queries;
+using Watchman.DomainModel.Protection;
+using Watchman.DomainModel.Protection.Commands;
+using Watchman.DomainModel.Protection.Queries;
 
 namespace Watchman.Discord.Areas.Protection.Services
 {
