+++ Watchman.DomainModel/Protection/Commands/AddProtectionPunishmentCommand.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Watchman.Cqrs;
 
-namespace Watchman.DomainModel.Users.Commands
+namespace Watchman.DomainModel.Protection.Commands
 {
     public class AddProtectionPunishmentCommand : ICommand
     {
