+++ Watchman.DomainModel/Protection/Queries/GetMuteEventsQuery.cs
@@ -1,6 +1,6 @@
 ﻿using Watchman.Cqrs;
 
-namespace Watchman.DomainModel.Users.Queries
+namespace Watchman.DomainModel.Protection.Queries
 {
     public class GetMuteEventsQuery : IQuery<GetMuteEventsQueryResult>
     {
