+++ Watchman.DomainModel/Protection/Queries/Handlers/GetProtectionPunishmentsQueryHandler.cs
@@ -1,7 +1,7 @@
 ﻿using Watchman.Cqrs;
 using Watchman.Integrations.MongoDB;
 
-namespace Watchman.DomainModel.Users.Queries.Handlers
+namespace Watchman.DomainModel.Protection.Queries.Handlers
 {
     public class GetProtectionPunishmentsQueryHandler : IQueryHandler<GetProtectionPunishmentsQuery, GetProtectionPunishmentsQueryResult>
     {
