--- Watchman.DomainModel/Users/Commands/Handlers/AddUserRoleCommandHandler.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Threading.Tasks;
-using Watchman.Cqrs;
-
-namespace Watchman.DomainModel.Users.Commands.Handlers
-{
-    public class AddUserRoleCommandHandler : ICommandHandler<AddUserRoleCommand>
-    {
-        public AddUserRoleCommandHandler()
-        {
-
-        }
-
-        public Task HandleAsync(AddUserRoleCommand command)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
