--- Watchman.DomainModel/Users/Queries/GetUserAvatarQuery.cs
@@ -1,8 +1,0 @@
-﻿using Watchman.Cqrs;
-
-namespace Watchman.DomainModel.Users.Queries
-{
-    public class GetUserAvatarQuery : IQuery<GetUserAvatarQueryResult>
-    {
-    }
-}
