--- Watchman.DomainModel/Users/Queries/GetUserAvatarQueryResult.cs
@@ -1,8 +1,0 @@
-﻿using Watchman.Cqrs;
-
-namespace Watchman.DomainModel.Users.Queries
-{
-    public class GetUserAvatarQueryResult : IQueryResult
-    {
-    }
-}
